+++ ImpactfulSkills/ImpactfulSkills.cs
@@ -24,7 +24,7 @@ namespace ImpactfulSkills
     {
         public const string PluginGUID = "MidnightsFX.ImpactfulSkills";
         public const string PluginName = "ImpactfulSkills";
-        public const string PluginVersion = "0.6.1";
+        public const string PluginVersion = "0.7.0";
 
         public ValConfig cfg;
         // Use this class to add your own localization to the game